+++ src/plugins/default/__tests__/index.js
@@ -1,9 +1,8 @@
+/* eslint-disable no-unused-expressions */
 import chai, { expect } from 'chai';
 import { countOccurences, runSSHCommand } from '../../../utils';
 import { describe, it } from 'mocha';
-
 import chaiString from 'chai-string';
-/* eslint-disable max-len, no-unused-expressions */
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
@@ -15,11 +14,11 @@ sh.config.silent = false;
 const servers = require('../../../../tests/fixtures/servers');
 
 describe('module - default', function() {
-  this.timeout(600000);
+  this.timeout(900000);
 
-  describe('deploy', function() {
+  describe('deploy', () => {
     it('should deploy meteor app on "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       sh.exec('mup setup');
 
@@ -27,11 +26,18 @@ describe('module - default', function() {
 
       expect(out.code).to.equal(0);
       expect(
-        countOccurences('Building App Bundle Locally', out.output)
-      ).to.be.equal(1);
+        out.output
+      ).satisfy(text => {
+        if (text.indexOf('Building App Bundle Locally') > -1) {
+          return true;
+        }
+
+        return text.indexOf('Using build from previous deploy at') > -1;
+      });
+
       expect(
         countOccurences(
-          'Pushing Meteor App Bundle to The Server: SUCCESS',
+          'Pushing Meteor App Bundle to the Server: SUCCESS',
           out.output
         )
       ).to.be.equal(1);
@@ -60,7 +66,7 @@ describe('module - default', function() {
     });
   });
 
-  describe('init', function() {
+  describe('init', () => {
     it('should create "mup.js" and "setting.json" in /tmp/project-tmp', () => {
       const dir = path.resolve(os.tmpdir(), 'project-tmp');
       sh.mkdir(dir);
@@ -72,7 +78,7 @@ describe('module - default', function() {
     });
   });
 
-  describe('logs', function() {
+  describe('logs', () => {
     it('should pull the logs from meteor app', () => {
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       const out = sh.exec('mup logs --tail 2');
@@ -80,9 +86,9 @@ describe('module - default', function() {
     });
   });
 
-  describe('reconfig', function() {
+  describe('reconfig', () => {
     it('should reconfig meteor app on "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       sh.exec('mup setup  && mup deploy --cached-build');
 
@@ -104,9 +110,9 @@ describe('module - default', function() {
     });
   });
 
-  describe('restart', function() {
+  describe('restart', () => {
     it('should restart meteor app on "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       sh.exec('mup setup  && mup deploy --cached-build');
 
@@ -124,9 +130,9 @@ describe('module - default', function() {
     });
   });
 
-  describe('setup', function() {
+  describe('setup', () => {
     it('should setup "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
 
       const out = sh.exec('mup setup');
@@ -147,9 +153,9 @@ describe('module - default', function() {
     });
   });
 
-  describe('start', function() {
+  describe('start', () => {
     it('should start meteor app on "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       sh.exec('mup setup  && mup meteor push --cached-build && mup meteor envconfig');
 
@@ -168,9 +174,9 @@ describe('module - default', function() {
     });
   });
 
-  describe('stop', function() {
+  describe('stop', () => {
     it('should stop meteor app on "meteor" vm', async () => {
-      const serverInfo = servers['mymeteor'];
+      const serverInfo = servers.mymeteor;
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-1'));
       sh.exec('mup setup  && mup deploy --cached-build');
 
@@ -186,10 +192,10 @@ describe('module - default', function() {
     });
   });
 
-  describe('syslog', function() {
-    const serverInfo = servers['mymeteor'];
+  describe('syslog', () => {
+    const serverInfo = servers.mymeteor;
 
-    it('should write meteor logst to syslog on "meteor" vm', async () => {
+    it('should write meteor logs to syslog on "meteor" vm', async () => {
       sh.cd(path.resolve(os.tmpdir(), 'tests/project-2'));
 
       sh.exec('mup setup && mup deploy --cached-build');