+++ src/plugins/meteor/build.js
@@ -1,7 +1,7 @@
-import tar from 'tar';
 import debug from 'debug';
 import fs from 'fs';
 import { spawn } from 'child_process';
+import tar from 'tar';
 
 const log = debug('mup:module:meteor');
 
@@ -10,7 +10,6 @@ export default function buildApp(appPath, buildOptions, verbose, api) {
   try {
     fs.statSync(api.resolvePath(appPath));
   } catch (e) {
-
     if (e.code === 'ENOENT') {
       console.log(`${api.resolvePath(appPath)} does not exist`);
     } else {
@@ -34,13 +33,15 @@ export default function buildApp(appPath, buildOptions, verbose, api) {
     const callback = err => {
       if (err) {
         reject(err);
+
         return;
       }
       resolve();
     };
-    buildMeteorApp(appPath, buildOptions, verbose, function(code) {
+    buildMeteorApp(appPath, buildOptions, verbose, code => {
       if (code === 0) {
         callback();
+
         return;
       }
       console.log('\n=> Build Error. Check the logs printed above.');
@@ -50,8 +51,8 @@ export default function buildApp(appPath, buildOptions, verbose, api) {
 }
 
 function buildMeteorApp(appPath, buildOptions, verbose, callback) {
-  var executable = buildOptions.executable || 'meteor';
-  var args = [
+  let executable = buildOptions.executable || 'meteor';
+  let args = [
     'build',
     '--directory',
     buildOptions.buildLocation,
@@ -72,7 +73,7 @@ function buildMeteorApp(appPath, buildOptions, verbose, callback) {
     args.push('--server-only');
   } else if (!buildOptions.mobileSettings) {
     args.push('--mobile-settings');
-    args.push(appPath + '/settings.json');
+    args.push(`${appPath}/settings.json`);
   }
 
   if (buildOptions.server) {
@@ -84,7 +85,7 @@ function buildMeteorApp(appPath, buildOptions, verbose, callback) {
     args.push('--allow-incompatible-update');
   }
 
-  var isWin = /^win/.test(process.platform);
+  const isWin = /^win/.test(process.platform);
   if (isWin) {
     // Sometimes cmd.exe not available in the path
     // See: http://goo.gl/ADmzoD
@@ -92,7 +93,7 @@ function buildMeteorApp(appPath, buildOptions, verbose, callback) {
     args = ['/c', 'meteor'].concat(args);
   }
 
-  var options = {
+  const options = {
     cwd: appPath,
     env: {
       ...process.env,
@@ -104,7 +105,7 @@ function buildMeteorApp(appPath, buildOptions, verbose, callback) {
   log(`Build Path: ${appPath}`);
   log(`Build Command:  ${executable} ${args.join(' ')}`);
 
-  var meteor = spawn(executable, args, options);
+  const meteor = spawn(executable, args, options);
 
   if (!verbose) {
     meteor.stdout.pipe(process.stdout, { end: false });