+++ src/plugins/mongo/command-handlers.js
@@ -1,3 +1,4 @@
+import chalk from 'chalk';
 import debug from 'debug';
 import nodemiral from 'nodemiral';
 
@@ -8,7 +9,10 @@ export function logs(api) {
 
   const args = api.getArgs();
   const sessions = api.getSessions(['mongo']);
-  args.shift(); // remove mongo from args sent to docker
+
+  // remove mongo from args sent to docker
+  args.shift();
+
   return api.getDockerLogs('mongodb', sessions, args);
 }
 
@@ -20,6 +24,7 @@ export function setup(api) {
     console.log(
       'Not setting up built-in mongodb since there is no mongo config'
     );
+
     return;
   }
 
@@ -30,11 +35,13 @@ export function setup(api) {
     console.log(
       'To use mup built-in mongodb setup, you should have only one meteor app server. To have more app servers, use an external mongodb setup'
     );
+
     return;
   } else if (mongoSessions[0]._host !== meteorSessions[0]._host) {
     console.log(
       'To use mup built-in mongodb setup, you should have both meteor app and mongodb on the same server'
     );
+
     return;
   }
 
@@ -66,6 +73,7 @@ export function start(api) {
     mongoSessions[0]._host !== meteorSessions[0]._host
   ) {
     log('Skipping mongodb start. Incompatible config');
+
     return;
   }
 
@@ -80,6 +88,7 @@ export function start(api) {
   });
 
   const sessions = api.getSessions(['mongo']);
+
   return api.runTaskList(list, sessions, { verbose: api.verbose });
 }
 
@@ -92,5 +101,93 @@ export function stop(api) {
   });
 
   const sessions = api.getSessions(['mongo']);
+
   return api.runTaskList(list, sessions, { verbose: api.verbose });
+}
+
+export async function status(api) {
+  const config = api.getConfig();
+
+  if (!config.mongo) {
+    return;
+  }
+  const mongoServer = Object.keys(config.mongo.servers)[0];
+  const server = config.servers[mongoServer];
+
+  const { output: dockerStatus } = await api.runSSHCommand(
+    server,
+    'docker inspect mongodb --format "{{json .}}"'
+  );
+  const mongoCommand = '"JSON.stringify(db.runCommand({serverStatus: 1, metrics: 0, wiredTiger: 1}))"';
+  let {
+    output: mongoStatus
+  } = await api.runSSHCommand(
+    server,
+    `docker exec mongodb mongo --eval ${mongoCommand} --quiet`
+  );
+
+  try {
+    mongoStatus = JSON.parse(mongoStatus);
+  } catch (e) {
+    console.log(chalk.red('\n=> Mongo Status'));
+    console.log(chalk.red(' - Stopped'));
+
+    return;
+  }
+
+  const mongoVersion = mongoStatus.version;
+  const connections = mongoStatus.connections.current;
+  const storageEngine = mongoStatus.storageEngine.name;
+
+  let containerStatus;
+  let statusColor = 'green';
+  let createdTime;
+  let restartCount = 0;
+  let restartCountColor = 'green';
+  let overallColor = 'green';
+
+  if (dockerStatus.trim() === '') {
+    containerStatus = 'Not started';
+    statusColor = 'red';
+  } else {
+    const info = JSON.parse(dockerStatus);
+    containerStatus = info.State.Status;
+
+    if (info.State.Restarting) {
+      statusColor = 'yellow';
+    } else if (info.State.Running !== true) {
+      statusColor = 'red';
+    }
+
+    const hour = 1000 * 60 * 60;
+    createdTime = info.Created;
+    const upTime = new Date(info.State.FinishedAt).getTime() -
+     new Date(info.Created).getTime();
+    restartCount = info.RestartCount;
+
+    if (restartCount > 0 && upTime / hour <= restartCount) {
+      restartCountColor = 'red';
+    } else if (restartCount > 1) {
+      restartCountColor = 'yellow';
+    }
+  }
+
+  if (
+    statusColor === 'green' &&
+    restartCountColor === 'green'
+  ) {
+    overallColor = 'green';
+  } else {
+    console.log('status', statusColor === 'green');
+    console.log('restart', restartCountColor === 'green');
+    overallColor = 'red';
+  }
+
+  console.log(chalk[overallColor]('\n=> Mongo Status'));
+  console.log(chalk[statusColor](`  ${containerStatus} on server ${server.host}`));
+  console.log(chalk[restartCountColor](`  Restarted ${restartCount} times`));
+  console.log(`  Running since ${createdTime}`);
+  console.log(`  Version: ${mongoVersion}`);
+  console.log(`  Connections: ${connections}`);
+  console.log(`  Storage Engine: ${storageEngine}`);
 }