+++ src/plugins/proxy/commands.js
@@ -42,3 +42,14 @@ export const stop = {
   description: 'Stop proxy',
   handler: commandHandlers.stop
 };
+
+export const nginxConfig = {
+  name: 'nginx-config',
+  description: 'View generated nginx config',
+  handler: commandHandlers.nginxConfig
+};
+
+export const status = {
+  description: 'View the proxy\'s status',
+  handler: commandHandlers.status
+};