+++ src/plugins/proxy/index.js
@@ -1,5 +1,4 @@
 import * as _commands from './commands';
-
 import validator from './validate';
 
 export const description = 'Setup and manage reverse proxy and ssl';
@@ -17,13 +16,35 @@ export function prepareConfig(config) {
 
   config.app.env = config.app.env || {};
 
-  config.app.env['VIRTUAL_HOST'] = config.proxy.domains;
-  config.app.env['HTTPS_METHOD'] = config.proxy.ssl && config.proxy.ssl.forceSSL ? 'redirect' : 'noredirect';
+  config.app.env.VIRTUAL_HOST = config.proxy.domains;
+  config.app.env.HTTPS_METHOD = config.proxy.ssl && config.proxy.ssl.forceSSL ? 'redirect' : 'noredirect';
+  config.app.env.HTTP_FORWARDED_COUNT =
+    config.app.env.HTTP_FORWARDED_COUNT || 1;
 
   if (config.proxy.ssl && config.proxy.ssl.letsEncryptEmail) {
-    config.app.env['LETSENCRYPT_HOST'] = config.proxy.domains;
-    config.app.env['LETSENCRYPT_EMAIL'] = config.proxy.ssl.letsEncryptEmail;
+    config.app.env.LETSENCRYPT_HOST = config.proxy.domains;
+    config.app.env.LETSENCRYPT_EMAIL = config.proxy.ssl.letsEncryptEmail;
   }
 
   return config;
+}
+
+export const hooks = {
+  'post.default.status'(api) {
+    if (api.getConfig().proxy) {
+      api.runCommand('proxy.status');
+    }
+  }
+};
+
+export function swarmOptions(config) {
+  if (config && config.proxy) {
+    const managers = [];
+    const servers = Object.keys(config.proxy.servers);
+    managers.push(servers[0]);
+
+    return {
+      managers
+    };
+  }
 }